backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
pgvector==0.2.4

# AI/LLM
anthropic==0.18.1
openai==1.12.0
tiktoken==0.5.2

# Communication
twilio==8.12.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.1

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.3
structlog==24.1.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Background Tasks
celery==5.3.6
